requirements.txt
# Backend Framework
fastapi
uvicorn[standard]
python-multipart

# Database
sqlalchemy
sqlmodel

# Frontend
streamlit

# APIs and Utilities
requests
tenacity
python-dotenv

# AI / RAG
chromadb
sentence-transformers
openai # Still needed for Ollama
azure-ai-inference # NEW: The correct SDK for GitHub Models

# Testing
pytest